robo3
Created components that get the card data, pass data as props to a cardList and card then render a list of cards on screen.  Added components and functionality for a searchbox and filtering the list or robots to display only those with the same name as what was searched.  Added scroll component that allows the user to scroll down through the robots.